requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
aiohttp>=3.9.0

# Security
pyjwt>=2.8.0
cryptography>=41.0.0
passlib>=1.7.4

# API and Web
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0
pydantic>=2.4.0

# Monitoring
prometheus-client>=0.17.0
grafana-api>=1.0.3
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.87.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
autodoc>=0.5.0

# Development tools
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0